Update 2018-01-08-200.3x-CICD-CDwithJenkins.md
@richieteh94 recommended updating the inline pipeline script for Jenkins CICD lab, Task 4, issue #153
The lab steps use an outdated URI on lines 192 and 261 in the lab markdown file 2018-01-08-200.3x-CICD-CDwithJenkins.md. The URI points to a script file used in the Jenkins pipeline for deploying artifacts to the Jenkins MRP app VM. The location of the script file has changed and the URI needs to be modified to match the new script location.