Update images to match current Azure UI
The deploy to Azure process involves more steps, and the existing screenshot images do not match the current Azure UI layout. The updated image files match the current Azure UI lay. The new image files are referenced, via HTML links, in the updated version of the 'Setup PU with Jenkins' lab on this branch.