Remove ref. to line num. from task4 pipline script
Task 4, "Update the Jenkins pipeline", provides instructions to paste copied code after line 30 of the pipeline script.
To avoid confusion, change the instruction to "Insert the following code before the final closing brace", rather than referring to a specific line number. The line spacing in the pipeline script from the prerequisite lab differs from the line spacing in the example pipeline script within this lab. As a result, the line numbers don't correlate.